refactor(api): 按端点拆分响应模型
